Add support for batched messages
`Statsd#batch_size` attribute sets the maximum number of messages to batch
before writing them to the UDP socket. The default setting of 1
maintains existing statsd behavior of sending all messages immediately

This introduces batching logic into the main control flow and may have
performance and/or thread-safety implications, but it is the simplest
thing that could possibly work and is hopefully a suitable proof of
concept.