Add the ability to pre-connect to the UDPSocket with `connect`
In testing with a previous version of the statsd-ruby gem we noticed
that a lot of time was spent sending these UDP messages. Looking
at the Ruby API, I noticed that by calling `connect` on the socket
ahead of time, future sends on the socket were significantly faster.

This adds a `connect` method, that when called, no longer needs to send
host/port with every send.